Add a Arrow VSI file system
Starting with GDAL 3.10 and Arrow 16.0, any GDAL Virtual File System can be
used (in a read-only context) wherever the Arrow C++ library expects a URI, in
particular outside of the context of the OGR Arrow driver, by:

- loading the libgdal.so/dll library (or the ogr_Arrow.so/dll plugin library if
  the Arrow driver is built as a library) with the arrow::fs::LoadFileSystemFactories()
  function (cf `Defining new filesystems <https://arrow.apache.org/docs/cpp/io.html#defining-new-filesystems>`__)
  Note: if the Arrow driver is fully loaded, e.g. by querying
  GetGDALDriverManager()->GetDriverByName("ARROW")->GetMetadata(), the Arrow VSI
  file system will be also registered.

- Prefixing any GDAL file name with the ``vsi://`` URI scheme prefix. In addition
  to any potential vsi prefix in the GDAL file name. So the ``/vsicurl/http://example.com``
  GDAL file name becomes the ``vsi:///vsicurl/http://example.com`` Arrow URI.
